+++ app/src/features/note/store/trash-api.ts
@@ -1,0 +1,7 @@
+import { notesApi } from "./notes-api";
+
+export const trashApi = notesApi.injectEndpoints({
+  endpoints: (builder) => ({}),
+});
+
+export const {} = trashApi;