+++ app/src/features/settings/store/settings-actions.ts
@@ -2,6 +2,7 @@ import { DarkwriteUserSettings } from "@/common/settings";
 import { DeepPartial } from "@/common/ts-util";
 import { store } from "@/features/store/redux";
 import { settingsSlice } from "./settings-slice";
+import { PageSize } from "@/common/pdf";
 
 export function updateSettings(partial: DeepPartial<DarkwriteUserSettings>) {
   store.dispatch(settingsSlice.actions.update(partial));
@@ -12,4 +13,12 @@ export function updateSettings(partial: DeepPartial<DarkwriteUserSettings>) {
  */
 export function getSettings() {
   return store.getState().settings;
+}
+
+export function setPreferredPageSize(size: PageSize) {
+  updateSettings({
+    editor: {
+      preferredPageSize: size,
+    },
+  });
 }